config for db, project structure refactor
we basically implemented a way to read from a yaml file and use it for configuration, this is a good practice for production apps, we also refactored the project structure to be more modular and easier to read. We now are also able to connect to the database via sqlx. GitHub CI also has been updated to reflect the changes.